average

+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -10,10 +10,10 @@ class StatsServiceTest {
     void shouldSumSales() {
         StatsService service = new StatsService();
 
-        int[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 180;
 
-        int actual = service.sumSales(enterpriseSales);
+        long actual = service.sumSales(enterpriseSales);
 
         assertEquals(expected, actual);
     }
@@ -22,10 +22,10 @@ class StatsServiceTest {
     void shouldaverageCompanySales() {
         StatsService service = new StatsService();
 
-        int[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 15;
 
-        int actual = service.averageCompanySales(enterpriseSales);
+        long actual = service.averageCompanySales(enterpriseSales);
 
         assertEquals(expected, actual);
     }
@@ -57,10 +57,10 @@ class StatsServiceTest {
     void shouldMonthsMinAverageCompanySales() {
         StatsService service = new StatsService();
 
-        int[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
 
-        int actual = service.monthsMinAverageCompanySales(enterpriseSales);
+        long actual = service.monthsMinAverageCompanySales(enterpriseSales);
 
         assertEquals(expected, actual);
     }
@@ -69,10 +69,10 @@ class StatsServiceTest {
     void shouldMonthsMaxAverageCompanySales() {
         StatsService service = new StatsService();
 
-        int[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] enterpriseSales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
 
-        int actual = service.monthMaxAverageCompanySales(enterpriseSales);
+        long actual = service.monthMaxAverageCompanySales(enterpriseSales);
 
         assertEquals(expected, actual);
     }
